put_comment_simplyを共通化
