+++ annofabcli/comment/put_inspection_comment_simply.py
@@ -1,23 +1,16 @@
 import argparse
 import logging
-import multiprocessing
 import sys
-import uuid
-from dataclasses import dataclass
-from functools import partial
-from typing import Any, Collection, Dict, List, Optional, Tuple
+from typing import Optional
 
-import annofabapi
-import annofabapi.utils
-from annofabapi.models import InputDataType, ProjectMemberRole, TaskPhase, TaskStatus
-from dataclasses_json import DataClassJsonMixin
+from annofabapi.models import CommentType, InputDataType, ProjectMemberRole
 
 import annofabcli
 import annofabcli.common.cli
+from annofabcli.comment.put_comment_simply import AddedSimpleComment, PutCommentSimplyMain
 from annofabcli.common.cli import (
     COMMAND_LINE_ERROR_STATUS_CODE,
     AbstractCommandLineInterface,
-    AbstractCommandLineWithConfirmInterface,
     ArgumentParser,
     build_annofabapi_resource_and_login,
     get_list_from_args,
@@ -25,190 +18,6 @@ from annofabcli.common.cli import (
 from annofabcli.common.facade import AnnofabApiFacade
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class AddedSimpleComment(DataClassJsonMixin):
-    """
-    付与するシンプルな検査コメント
-    """
-
-    comment: str
-    """検査コメントの中身"""
-
-    data: Dict[str, Any]
-    """検査コメントを付与する位置や区間"""
-
-    phrases: Optional[List[str]] = None
-    """参照している定型指摘ID"""
-
-
-class PutInspectionCommentSimplyMain(AbstractCommandLineWithConfirmInterface):
-    def __init__(self, service: annofabapi.Resource, project_id: str, all_yes: bool = False):
-        self.service = service
-        self.facade = AnnofabApiFacade(service)
-        self.project_id = project_id
-
-        AbstractCommandLineWithConfirmInterface.__init__(self, all_yes)
-
-    def _create_request_body(self, task: Dict[str, Any], comment_info: AddedSimpleComment) -> List[Dict[str, Any]]:
-        """batch_update_comments に渡すリクエストボディを作成する。"""
-
-        def _convert(comment: AddedSimpleComment) -> Dict[str, Any]:
-            return {
-                "comment": comment.comment,
-                "comment_id": str(uuid.uuid4()),
-                "phase": task["phase"],
-                "phase_stage": task["phase_stage"],
-                "comment_type": "inspection",
-                "account_id": self.service.api.account_id,
-                "comment_node": {"data": comment.data, "status": "open", "_type": "Root"},
-                "phrases": comment.phrases,
-                "_type": "Put",
-            }
-
-        return [_convert(comment_info)]
-
-    def change_to_working_status(self, project_id: str, task: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        作業中状態に遷移する。必要ならば担当者を自分自身に変更する。
-
-        Args:
-            project_id:
-            task:
-
-        Returns:
-            作業中状態遷移後のタスク
-        """
-
-        task_id = task["task_id"]
-
-        if task["account_id"] != self.service.api.account_id:
-            self.service.wrapper.change_task_operator(project_id, task_id, self.service.api.account_id)
-            logger.debug(f"{task_id}: 担当者を自分自身に変更しました。")
-
-        changed_task = self.service.wrapper.change_task_status_to_working(project_id, task_id)
-        return changed_task
-
-    @staticmethod
-    def _can_add_comment(
-        task: Dict[str, Any],
-    ) -> bool:
-        task_id = task["task_id"]
-        if task["phase"] == TaskPhase.ANNOTATION.value:
-            logger.warning(f"task_id='{task_id}': 教師付フェーズなので、検査コメントを付与できません。")
-            return False
-
-        if task["status"] not in [TaskStatus.NOT_STARTED.value, TaskStatus.WORKING.value, TaskStatus.BREAK.value]:
-            logger.warning(
-                f"task_id='{task_id}' : タスクの状態が未着手,作業中,休憩中 以外の状態なので、検査コメントを付与できません。（task_status='{task['status']}'）"
-            )
-            return False
-        return True
-
-    def put_inspection_comment_for_task(
-        self,
-        task_id: str,
-        comment_info: AddedSimpleComment,
-        task_index: Optional[int] = None,
-    ) -> bool:
-        """
-        タスクに検査コメントを付与します。
-
-        Args:
-            task_id: タスクID
-            comment_info: コメント情報
-            task_index: タスクの連番
-
-        Returns:
-            付与した検査コメントの数
-        """
-        logging_prefix = f"{task_index+1} 件目" if task_index is not None else ""
-
-        task = self.service.wrapper.get_task_or_none(self.project_id, task_id)
-        if task is None:
-            logger.warning(f"{logging_prefix} : task_id='{task_id}' のタスクは存在しないので、スキップします。")
-            return False
-
-        logger.debug(
-            f"{logging_prefix} : task_id = {task['task_id']}, "
-            f"status = {task['status']}, "
-            f"phase = {task['phase']}, "
-        )
-
-        if not self._can_add_comment(
-            task=task,
-        ):
-            return False
-
-        if not self.confirm_processing(f"task_id='{task_id}' のタスクに検査コメントを付与しますか？"):
-            return False
-
-        # 検査コメントを付与するには作業中状態にする必要があるので、タスクの状態を作業中にする
-        changed_task = self.change_to_working_status(self.project_id, task)
-
-        input_data_id = task["input_data_id_list"][0]
-
-        try:
-            # 検査コメントを付与する
-            request_body = self._create_request_body(task=changed_task, comment_info=comment_info)
-            self.service.api.batch_update_comments(self.project_id, task_id, input_data_id, request_body=request_body)
-            logger.debug(f"{logging_prefix} : task_id={task_id} のタスクに検査コメントを付与しました。")
-            return True
-        except Exception:  # pylint: disable=broad-except
-            logger.warning(
-                f"{logging_prefix} : task_id={task_id}, input_data_id={input_data_id}: 検査コメントの付与に失敗しました。", exc_info=True
-            )
-            return False
-        finally:
-            self.service.wrapper.change_task_status_to_break(self.project_id, task_id)
-            # 担当者が変えている場合は、元に戻す
-            if task["account_id"] != changed_task["account_id"]:
-                self.service.wrapper.change_task_operator(self.project_id, task_id, task["account_id"])
-                logger.debug(f"{task_id}: 担当者を元のユーザ( account_id={task['account_id']}）に戻しました。")
-
-    def add_comments_for_task_wrapper(self, tpl: Tuple[int, str], comment_info: AddedSimpleComment) -> bool:
-        task_index, task_id = tpl
-        try:
-            return self.put_inspection_comment_for_task(
-                task_id=task_id, comment_info=comment_info, task_index=task_index
-            )
-        except Exception:  # pylint: disable=broad-except
-            logger.warning(f"task_id={task_id}: 検査コメントの付与に失敗しました。", exc_info=True)
-            return False
-
-    def put_inspection_comment_for_task_list(
-        self,
-        task_ids: Collection[str],
-        comment_info: AddedSimpleComment,
-        parallelism: Optional[int] = None,
-    ) -> None:
-
-        logger.info(f"{len(task_ids)} 件のタスクに検査コメントを付与します。")
-
-        if parallelism is not None:
-            func = partial(self.add_comments_for_task_wrapper, comment_info=comment_info)
-            with multiprocessing.Pool(parallelism) as pool:
-                result_bool_list = pool.map(func, enumerate(task_ids))
-                success_count = len([e for e in result_bool_list if e])
-
-        else:
-            # 逐次処理
-            success_count = 0
-            for task_index, task_id in enumerate(task_ids):
-                try:
-                    result = self.put_inspection_comment_for_task(
-                        task_id=task_id,
-                        comment_info=comment_info,
-                        task_index=task_index,
-                    )
-                    if result:
-                        success_count += 1
-                except Exception:  # pylint: disable=broad-except
-                    logger.warning(f"task_id={task_id}: 検査コメントの付与に失敗しました。", exc_info=True)
-                    continue
-
-        logger.info(f"{success_count} / {len(task_ids)} 件のタスクに検査コメントを付与しました。")
 
 
 class PutInspectionCommentSimply(AbstractCommandLineInterface):
@@ -250,8 +59,10 @@ class PutInspectionCommentSimply(AbstractCommandLineInterface):
 
         task_id_list = get_list_from_args(args.task_id)
         phrase_id_list = get_list_from_args(args.phrase_id)
-        main_obj = PutInspectionCommentSimplyMain(self.service, project_id=args.project_id, all_yes=self.all_yes)
-        main_obj.put_inspection_comment_for_task_list(
+        main_obj = PutCommentSimplyMain(
+            self.service, project_id=args.project_id, comment_type=CommentType.INSPECTION, all_yes=self.all_yes
+        )
+        main_obj.put_comment_for_task_list(
             task_ids=task_id_list,
             comment_info=AddedSimpleComment(comment=args.comment, data=comment_data, phrases=phrase_id_list),
             parallelism=args.parallelism,
