+++ annofabcli/common/enums.py
@@ -41,3 +41,10 @@ class FormatArgument(Enum):
 
     #: project_idの一覧
     PROJECT_ID_LIST = "project_id_list"
+
+
+class CustomProjectType(Enum):
+    """カスタムプロジェクトの場合、検査コメントのフォーマットが分からないため、カスタムプロジェクトの種類をannofabcliで定義する。"""
+
+    THREE_DIMENSION_POINT_CLOUD = "3dpc"
+    """3DPCプロジェクト"""
